Correct double-mouse-click error ================================
1.  Modify glui.g so that variable last_click_time is now
    a 'long int'.

+++ include/GL/glui.h
@@ -2128,7 +2128,7 @@ public:
     GLUI_CB             obj_cb;
     int                 cb_click_type;
     int                 last_line;
-    int                 last_click_time;
+    long int            last_click_time;        /* Time of last mouse click */
 
     int  mouse_down_handler( int local_x, int local_y );
     int  mouse_up_handler( int local_x, int local_y, bool inside );
